Create Semantic Tags in HTML
Semantics are used to give proper meaning to the websites and help search engines to analyze it properly. Also, we can get help in return from different search engines in recognizing the original tags and their uses. For example, there are different tags used for the headers, navigation bars, body, and footers. So to analyze each and every tag properly, we have to use Semantics. The better the use of Semantics in a website, the more is the number of chances that it is being crawled by the search engines. It will also help in the better ranking of a website.

In very simple language, Semantic means to provide meaning to any word. If we talk in about a web page, then earlier most of the websites are made only using divs including different classes and IDs in them. But that was not the correct way of telling the search engines that if the elements are header, footer, or any other body.

There are many examples of Semantics. You can take the reference of the internet to view all. Here are some of the important semantic elements-

<header>
<nav>
<section>
<article>
<footer>